% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4romp-package.R
\name{cpp4romp-package}
\alias{cpp4romp-package}
\title{Demonstration that 'cpp4r' can Use OpenMP with the Right Compiler}
\description{
The goal of this package is to demonstrate some features of the 
 cpp4r package, and this also demonstrates how to create named and unnamed lists.
 Build this package with gcc or clang on any computer with an Intel 
 or AMD processor to compile the contained C++ code with OpenMP support. It also
 works if you install libomp via Homebrew on macOS.
 The Makevars file is set up to use the OpenMP flag for the compiler, and 
 commenting its lines generates a build that works, but the functions will 
 use one thread.
}

