% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4romp-package.R
\name{squared_unnamed}
\alias{squared_unnamed}
\title{Unnamed list with squared numbers and the threads used}
\usage{
squared_unnamed(x)
}

\arguments{
\item{x}{A vector of doubles}
}

\description{
Unnamed list with squared numbers and the threads used
}

